Refactoring SolidLogin - WIP
